backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.31
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
minio==7.2.7
scikit-learn
numpy
PyPDF2
python-docx
pydantic==2.8.2
email-validator==2.2.0
python-dotenv==1.1.1
requests==2.32.3
sentence-transformers
torch
transformers
huggingface-hub
